Implements Edges, Generate Edges, and Sort Edges
Edges are what will store data for the Add Shortest alg to know what order to look at adding connections. Generate Edges is a helper function within Add Shortest designed to generate this list, and Sort Edges is used to generate the correct order of edges to be looked at.