Implements data runner (#11)
This is the bread and butter of the entire project. This change enables us to peek into the runtime and results of each algorithm. Everything is pulled together in this change.

Bug fixed: The way we were attaching nodes was incorrectly setting the opposite end on nodes. Because of this, we were adding too many connections via AS alg. Added additional testing to prevent regression in the future.